mcp-core: hoist duplicated chain_scc_project_json test helper
The read_model and edit_model E2E suites each carried their own copy of the
~40-line chain_scc_project_json fixture (the 51-node single-SCC model that trips
the LTM auto-flip-to-discovery gate). The two copies had drifted: edit_model's
simSpecs included saveStep/timeUnits and read_model's did not.

Hoist one canonical copy into test_support.rs (the shared integration helper
module for this crate's single-harness convention) and import it from both
suites. Keep the fuller simSpecs (saveStep/timeUnits) as the canonical shape;
both suites only assert on the auto-flip warning, so the extra fields are inert.
