+++ src/simlin-mcp-core/tests/integration/read_model_e2e.rs
@@ -8,54 +8,8 @@
 //! `ProjectAccess` impl.
 
 use simlin_mcp_core::errors::AccessError;
-use simlin_mcp_core::test_support::TestFileSystemAccess;
+use simlin_mcp_core::test_support::{TestFileSystemAccess, chain_scc_project_json};
 use simlin_mcp_core::tools::read_model::{ReadModelInput, read_model};
-
-/// Build a native-JSON project whose causal graph has a single SCC of
-/// `total_nodes` nodes (a stock, a flow, and `total_nodes - 2` chained
-/// auxes), which exceeds the engine's `MAX_LTM_SCC_NODES = 50` auto-flip
-/// threshold when `total_nodes >= 51`. With LTM enabled, compiling this model
-/// emits the "auto-switched ... to discovery mode" Warning diagnostic.
-///
-/// Chain: `cap_stock -> aux_{N-3} -> ... -> aux_0 -> cap_flow -> cap_stock`.
-fn chain_scc_project_json(total_nodes: usize) -> serde_json::Value {
-    let aux_count = total_nodes - 2;
-    let auxiliaries: Vec<serde_json::Value> = (0..aux_count)
-        .map(|i| {
-            let equation = if i + 1 == aux_count {
-                "cap_stock".to_string()
-            } else {
-                format!("aux_{}", i + 1)
-            };
-            serde_json::json!({
-                "uid": (i + 10) as i64,
-                "name": format!("aux_{i}"),
-                "equation": equation,
-            })
-        })
-        .collect();
-
-    serde_json::json!({
-        "name": "chain_scc",
-        "simSpecs": {
-            "startTime": 0.0,
-            "endTime": 10.0,
-            "dt": "1",
-            "method": "euler"
-        },
-        "models": [{
-            "name": "main",
-            "stocks": [
-                {"uid": 1, "name": "cap_stock", "initialEquation": "0",
-                 "inflows": ["cap_flow"], "outflows": []}
-            ],
-            "flows": [
-                {"uid": 2, "name": "cap_flow", "equation": "aux_0"}
-            ],
-            "auxiliaries": auxiliaries
-        }]
-    })
-}
 
 #[tokio::test]
 async fn read_model_returns_clean_xmile_snapshot() {
